Add missing endpoints (#16)
* Add dependency to @Nullable annotation

* Remove plugin system

* Add nullable to some fields

* Add bouncer endpoints

* Add update profile endpoint

* Make some parameters optional

* Add countries and states

* Add timezones

* Make more parameters optional

* Return maps

* Updating donations needs user update token

* Return service id when ordering

* Add update status

* Add auto extension duration

* Test for null in service properties

* Add @Nullable to Gameserver and Cloudserver properties

* Make ProductionHttpClient extendable

* Test game for null

* Add detailed profile

* Add rescue mode

+++ src/test/java/net/nitrado/api/test/NitrapiTest.java
@@ -22,10 +22,10 @@ public class NitrapiTest extends ApiTestCase {
     public void testCustomer() throws NitrapiException {
         client.addNextResponseFromFile("customer/customer.json");
         Customer customer = api.getCustomer();
-        assertEquals(1337, customer.getUserId());
+        assertEquals(1337, (int)customer.getUserId());
         assertEquals("Marty", customer.getUsername());
         assertEquals("America/Los_Angeles", customer.getTimezone());
-        assertEquals(1955, customer.getCredit());
+        assertEquals(1955, (int)customer.getCredit());
         assertEquals("marty.mcfly@biffco.com", customer.getEmail());
         assertEquals("https://server.nitrado.net/users.nitrado/1337.jpg", customer.getAvatar());
         assertDate(1985, 10, 25, 01, 15, 00, customer.getRegistered());
